requirements.txt
# System
PyQt5
PyQt-Fluent-Widgets[full]
openai
anthropic
google-generativeai
tiktoken
# opencc
openpyxl
# ebooklib
# beautifulsoup4

# Tools
rich
tqdm
python-rapidjson